Working with Core Comment System + Disqus

+++ routes/web/posts.php
@@ -2,13 +2,17 @@
 
 use Illuminate\Support\Facades\Route;
 
-Route::get('/posts', 'PostController@index')->name('posts.index');
-Route::get('/posts/create', 'PostController@create')->name('post.create');
-Route::post('/posts/store', 'PostController@store')->name('post.store');
 Route::get('/post/{post}', 'PostController@show')->name('post');
-// Route::get('/posts/{post}/edit', 'PostController@edit')->name('post.edit');
-Route::put('/posts/{post}/update', 'PostController@update')->name('post.update');
-Route::delete('/posts/{post}/delete', 'PostController@delete')->name('post.delete');
+
+Route::middleware('auth')->group(function(){
+Route::get('admin/posts', 'PostController@index')->name('posts.index');
+
+        Route::get('admin/posts/create', 'PostController@create')->name('post.create');
+        Route::post('admin/posts/store', 'PostController@store')->name('post.store');
+        // Route::get('/posts/{post}/edit', 'PostController@edit')->name('post.edit');
+        Route::put('admin/posts/{post}/update', 'PostController@update')->name('post.update');
+        Route::delete('admin/posts/{post}/delete', 'PostController@delete')->name('post.delete');
+});
 
 
 // Route::get('/posts', 'PostController@index')->name('posts.index');
